redirects indicator tests to BehaviorOphysMetadata testing
